requirement.txt
apache-beam==2.22.0
avro-python3==1.9.2.1
bcrypt==3.1.7
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
crcmod==1.7
cryptography==2.9.2
dataclasses==0.7
dill==0.3.1.1
docopt==0.6.2
fastavro==0.23.6
fasteners==0.15
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.10.1
google-apitools==0.5.31
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-cloud-bigquery==1.24.0
google-cloud-bigtable==1.0.0
google-cloud-core==1.3.0
google-cloud-datastore==1.7.4
google-cloud-dlp==0.13.0
google-cloud-language==1.3.0
google-cloud-pubsub==2.0.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.13.0
google-cloud-storage==1.29.0
google-cloud-videointelligence==1.13.0
google-cloud-vision==0.42.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.30.0
grpcio-gcp==0.2.2
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
libcst==0.3.10
mock==2.0.0
monotonic==1.5
mypy-extensions==0.4.3
numpy==1.19.0
oauth2client==3.0.0
paramiko==2.7.1
pbr==5.4.5
proto-plus==1.9.1
protobuf==3.12.2
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pymongo==3.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
storage==0.0.4.3
typing-extensions==3.7.4.2
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
